Check if directory exists, before we remove it
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10999 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ build/jenkins/clean.py
@@ -4,5 +4,6 @@ from util import *
 #
 # Clean up the build directory. Can be used to ensure a completely clean build.
 #
-shutil.rmtree(builddir)
+if os.path.exists(builddir):
+  shutil.rmtree(builddir)
